move OphysROI to ophys_etl/types/_roi.py
