+++ src/ophys_etl/modules/decrosstalk/qc_plotting/summary_plot.py
@@ -16,8 +16,8 @@ import scipy.stats
 
 from typing import Tuple, List
 
+from ophys_etl.types import OphysROI
 from ophys_etl.modules.decrosstalk.ophys_plane import DecrosstalkingOphysPlane
-from ophys_etl.modules.decrosstalk.ophys_plane import OphysROI
 from ophys_etl.modules.decrosstalk.qc_plotting.utils import add_gridlines
