+++ src/ophys_etl/types/__init__.py
@@ -1,1 +1,1 @@
-from ophys_etl.types._roi import DenseROI, ExtractROI  # noqa: F401
+from ophys_etl.types._roi import DenseROI, ExtractROI, OphysROI  # noqa: F401
