+++ tests/modules/decrosstalk/test_ophys_plane.py
@@ -1,6 +1,6 @@
 import os
 import json
-from ophys_etl.modules.decrosstalk.ophys_plane import OphysROI
+from ophys_etl.types import OphysROI
 from ophys_etl.modules.decrosstalk.ophys_plane import DecrosstalkingOphysPlane
 
 from .utils import get_data_dir
