+++ tests/modules/decrosstalk/test_qc_plotting.py
@@ -8,7 +8,7 @@ import h5py
 import numpy as np
 import PIL
 
-from ophys_etl.modules.decrosstalk.ophys_plane import OphysROI
+from ophys_etl.types import OphysROI
 from ophys_etl.modules.decrosstalk.ophys_plane import DecrosstalkingOphysPlane
 
 from ophys_etl.modules.decrosstalk.qc_plotting.pairwise_plot import (
